api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
slack-sdk==3.27.1
python-multipart==0.0.9
